refactoer validatioon for all text input
