+++ src/data/constants/inputType.ts
@@ -1,9 +1,8 @@
 export const InputConstants = {
-  USERNAME: "username",
   PASSWORD: "password",
   EMAIL: "email",
-  FIRSTNAME: "firstname",
-  LASTNAME: "lastname",
   CONFIRM_PASSWORD: "confirmPassword",
   FULL_NAME: "fullname",
+  COMPANY_NAME: "companyName",
+  PHONE_NUMBER: "phoneNumber",
 };
