+++ src/features/recruiter/authentication/index.ts
@@ -1,0 +1,1 @@
+export { default as RecruiterRegisterForm } from "./components/RecruiterRegisterForm";
